TCP reset forgery now work roughly.  There's no way to query the context to check your resetting the right thing, so bit of a blunt tool at the moment.  Also, hard-coded reset on the loopback port.  More work to do.

+++ ip.C
@@ -286,7 +286,7 @@ uint16_t ip::calculate_cksum(pdu_iter s, pdu_iter e)
     
     pdu_iter ptr = s;
 
-    unsigned long sum = 0;
+    uint32_t sum = 0;
 
     // Handle 2-bytes at a time.
     while ((e - ptr) > 1) {
